requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.32.3
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.0
